Replace remote test with unit test
