+++ package.json
@@ -69,9 +69,10 @@
         "lint-js": "standard --verbose | snazzy",
         "lint": "npm-run-all lint-js",
         "postinstall": "node scripts/bundle.js",
+        "test-unit": "cross-env NODE_PATH=lib:test ava test/unit",
         "test-com": "cross-env NODE_PATH=lib:test ava test/com",
         "test-int": "cross-env NODE_PATH=lib:test ava test/int",
-        "test": "npm-run-all test-com test-int",
+        "test": "npm-run-all test-unit test-com test-int",
         "test-hoc": "cross-env NODE_PATH=lib:test ava"
     }
 }
